api-request/pyproject.toml
[project]
name = "api-request"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "kafka-python"
]
